backend/requirements.txt
# Core FastAPI and Web Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
pydantic>=2.4.0
python-dotenv>=1.0.0

# AI and Machine Learning
sentence-transformers>=2.2.2
faiss-cpu>=1.7.4
numpy>=1.26.0
torch>=2.0.0
transformers>=4.30.0

# Authentication and Security
bcrypt>=4.0.1
PyJWT>=2.8.0
email-validator>=2.1.0

# File Processing
PyPDF2>=3.0.1
Pillow>=10.0.0
python-docx>=1.0.0
openpyxl>=3.1.0
pytesseract>=0.3.10

# Web Scraping and Data Processing
requests>=2.31.0
beautifulsoup4>=4.12.2
lxml>=4.9.0

# Database and Storage
# sqlite3 is built into Python

# Utilities
typing-extensions>=4.5.0